cleanup imports, move embeddings to modules
