+++ metaseq/modules/transformer_decoder_layer.py
@@ -10,15 +10,11 @@ import torch.nn as nn
 from torch import Tensor
 
 from metaseq import utils
-from metaseq.modules import gelu, MultiheadAttention
-from metaseq.modules.dropout import Dropout
-from metaseq.modules.feedforward_network import FeedForwardNetwork
+from metaseq.modules import gelu, MultiheadAttention, Dropout, FeedForwardNetwork, LayerNorm, Linear
 from metaseq.modules.fused_bias_gelu import (
     has_fused_bias_gelu,
     load_megatron_fused_kernel,
 )
-from metaseq.modules.layer_norm import LayerNorm
-from metaseq.modules.linear import Linear
 
 
 class TransformerDecoderLayer(nn.Module):
